refactor: optimize types
