+++ src/types/skills.ts
@@ -1,0 +1,12 @@
+import type { User } from './users'
+
+export type Skill = {
+  id: string
+  name: string
+  users: User[]
+}
+
+export type SkillCategory = {
+  name: string
+  skills: Skill[]
+}
