+++ src/types/users.ts
@@ -1,0 +1,5 @@
+export type User = {
+  id: string
+  name: string
+  image: string
+}
